# openapi.yaml
openapi: 3.0.3
info:
  title: StitchAdmin 2.0 API
  description: |
    Modernisierte ERP-Lösung für Stickerei-Betriebe

    ## Authentifizierung
    Die meisten Endpunkte erfordern eine Session-basierte Authentifizierung über Flask-Login.
    Melden Sie sich zuerst über `/auth/login` an, um eine Session zu erhalten.

    ## Berechtigung
    Viele Endpunkte erfordern Admin-Rechte. Diese sind mit dem Tag `admin` gekennzeichnet.

    ## Datenformate
    - Alle Datumsangaben sind im ISO 8601 Format (YYYY-MM-DD oder YYYY-MM-DDTHH:MM:SS)
    - Preise sind als Float-Werte in Euro
    - IDs sind alphanumerische Strings oder Integers je nach Entität

    Erstellt von Hans Hahn - Alle Rechte vorbehalten
  version: 2.0.0
  contact:
    name: StitchAdmin Support
    email: admin@stitchadmin.local
servers:
  - url: http://localhost:5000
    description: Entwicklungsserver
  - url: https://stitchadmin.example.com
    description: Produktionsserver
...
components:
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: session
      description: Session-Cookie nach Login
  schemas:
    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          example: "Fehlermeldung"
      required:
        - success
        - error
...
    Article:
      type: object
      properties:
        id:
          type: string
          example: "ART001"
        article_number:
          type: string
          example: "A-2025-001"
        name:
          type: string
          example: "T-Shirt Basic Weiß"
        description:
          type: string
          example: "Hochwertiges Baumwoll-T-Shirt"
        supplier:
          type: string
          example: "L-Shop"
        supplier_article_number:
          type: string
          example: "TS-001-W"
        category:
          type: string
          example: "Textilien"
        brand:
          type: string
          example: "Fruit of the Loom"
        price:
          type: number
          format: float
          example: 15.99
          description: "Verkaufspreis"
        purchase_price_single:
          type: number
          format: float
          example: 8.50
        purchase_price_carton:
          type: number
          format: float
          example: 7.20
        purchase_price_10carton:
          type: number
          format: float
          example: 6.80
        stock:
          type: integer
          example: 50
        min_stock:
          type: integer
          example: 10
        unit:
          type: string
          example: "Stück"
        barcode:
          type: string
          example: "4260123456789"
        active:
          type: boolean
          example: true
        created_at:
          type: string
          format: date-time
...
security:
  - sessionAuth: []
paths:
...
  /api/articles:
    get:
      tags:
        - API
      summary: Artikel-Liste abrufen
      description: Holt Artikel mit optionalen Filtern für Lieferant und Suche
      parameters:
        - name: supplier_id
          in: query
          schema:
            type: string
          description: Lieferanten-ID zum Filtern
        - name: search
          in: query
          schema:
            type: string
          description: Suchbegriff (Artikelnummer, Name, Lieferanten-Artikelnummer)
      responses:
        '200':
          description: Erfolgreiche Abfrage
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  articles:
                    type: array
                    items:
                      $ref: '#/components/schemas/Article'
                  count:
                    type: integer
                    example: 42
                  source:
                    type: string
                    enum: [database, json_fallback]
                    example: "database"
        '500':
          description: Serverfehler
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/articles/{article_id}:
    get:
      tags:
        - API
      summary: Artikel-Details abrufen
      description: Holt Details eines spezifischen Artikels
      parameters:
        - name: article_id
          in: path
          required: true
          schema:
            type: string
          description: Artikel-ID
      responses:
        '200':
          description: Erfolgreiche Abfrage
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  article:
                    $ref: '#/components/schemas/Article'
        '404':
          description: Artikel nicht gefunden
        '500':
          description: Serverfehler
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/articles/search:
    get:
      tags:
        - API
      summary: Artikel-Suche für Kassensystem
      description: Sucht Artikel nach Artikelnummer, Name oder Barcode
      parameters:
        - name: q
          in: query
          schema:
            type: string
          description: Suchbegriff
      responses:
        '200':
          description: Suchergebnisse
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: string
                    article_number:
                      type: string
                    name:
                      type: string
                    price:
                      type: number
                      format: float
                    stock_quantity:
                      type: integer
                    barcode:
                      type: string
                    category:
                      type: string
  /api/top-selling-articles:
    get:
      tags:
        - API
      summary: Meistverkaufte Artikel
      description: Holt die am häufigsten verkauften Artikel für Schnellzugriff
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 6
          description: Anzahl der Artikel
      responses:
        '200':
          description: Top-Artikel
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: string
                    article_number:
                      type: string
                    name:
                      type: string
                    price:
                      type: number
                    stock_quantity:
                      type: integer

  # ==================== API - DEBUG ====================
...
  /articles:
    get:
      tags:
        - Artikel
      summary: Artikel-Übersicht
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: category
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Artikel-Liste HTML
...
  /articles/{article_id}:
    get:
      tags:
        - Artikel
      summary: Artikel-Details
      parameters:
        - name: article_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Artikel-Details HTML
...
  /articles/{article_id}/stock:
    post:
      tags:
        - Artikel
      summary: Lagerbestand aktualisieren
      parameters:
        - name: article_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                action:
                  type: string
                  enum: [add, remove, set]
                quantity:
                  type: integer
                order_number:
                  type: string
                reason:
                  type: string
      responses:
        '302':
          description: Redirect zurück
...